move environ to conftest
